Use the golangci-lint GitHub Action (#401)
It seems that there's a change in golangci-lint that makes CI runs fail.

https://github.com/aviator-co/av/pull/399

Use the official GitHub Action for golangci-lint instead of installing
it manually.

https://github.com/aviator-co/av/issues/402

+++ internal/gh/pullrequest.go
@@ -88,7 +88,7 @@ type GetPullRequestsInput struct {
 	HeadRefName string
 	BaseRefName string
 	States      []githubv4.PullRequestState
-	First       int64
+	First       int32
 	After       string
 }
 
@@ -210,7 +210,7 @@ func (c *Client) MarkPullRequestReadyForReview(
 type RepoPullRequestOpts struct {
 	Owner  string
 	Repo   string
-	First  int64
+	First  int32
 	After  string
 	States []githubv4.PullRequestState
 }
